requirements.txt
# FastAPI ve Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning
# NOTE: `ml/models/mobilenetv2_final.keras` was saved with newer Keras serialization.
# TensorFlow 2.16+ bundles Keras 3-compatible loading.
tensorflow==2.16.1
numpy==1.24.3
Pillow==10.1.0

# Firebase ve Google Cloud
firebase-admin==6.2.0
google-cloud-firestore==2.13.1

# AI Chat
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# WebSocket
websockets==12.0
